requirements.txt
asgiref==3.8.1
Django==5.1.7
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.1.0